Notification Updates
Site-Wide Notifications and Course Smart Alerts are now dismissable. This information is saved to the user’s session, so alerts will be displayed again the next time a user logs in. Course alerts are dismissed on a per-alert, per-user, per-course, per-session basis, i.e. that particular alert will be hidden for that user on that course, but only until they log out.